+++ python_modules/dagster/dagster_tests/execution_tests/versioning_tests/test_data_versions.py
@@ -1033,6 +1033,6 @@ def test_fan_in():
         traced_counter.get().counts()  # pyright: ignore[reportOptionalMemberAccess]
         == {
             "DagsterInstance.get_asset_records": 1,
-            "DagsterInstance.get_run_record_by_id": 3,  # get_run_record_by_id called when handling events for the run
+            "RunsMixin.get_run_record_by_id": 3,  # get_run_record_by_id called when handling events for the run
         }
     )
